Remove unnecessary use of `_$bundler_version\_`.
Our build scripts used this to ensure that the expected version of bundler
(e.g. the version recorded as `BUNDLED WITH` in the `Gemfile.lock`) is used.
While there was a time this was necessary and useful, modern versions of
bundler do this automatically. If you run a `bundle` command with a different
version of bundler compared to what is recorded under `BUNDLED WITH`, it'll
install the expected version of bundler and re-run the command with it. We
simply no longer need this and it adds complexity.