Fix initial type of `Index#routing_field_path`.
This is a followup on #94. For historical/vestigial reasons, we
used to set `routing_field_path` to `[]` before immediately updating
it to the `FieldPath` object for `id`. With our change in #114 to
lazily set `routing_field_path`, it allowed the value of `[]` to
persists beyond `initialize`. `[]` isn't a valid value according
to the type; instead I've changed the type to allow `nil` which is
a more natural value for this case.