Allow query adapters to be instantiated without a schema.
The refactoring I'm working on requires that the query adapters
be built before the schema instance is available. As such, we won't
be able to pass it into the constructors. To allow for this, I've
made some simple `WithoutSchema` classes that we instantiate up
front, and then defined `call` to build the original classes using
the `elastic_graph_schema` available in `context`.