Optimize: avoid unneeded id-only datastore query.
On an nested relationship that uses an outbound foreign key,
we already have the `id` values before resolving the nested
relationship. While we must resolve the relationship in most
cases, we're able to just return the ids we already have so
long as the query satisfies the following crtieria:

* The query is only requesting `id`. If any other fields are requested
  we (obviously) need to execute the datastore query.
* The query is sorting on `id`, if there is more than one. If we have
  multiple ids and we're sorting on another field, we have to execute
  the datastore query to get the results in the requested order.

This optimization avoids an unnecessary datastore query for each
parent document, and should make a big efficiency/performance
difference in this case.

Note: while this should generally have no impact on the response,
it will have an impact when we have orphaned foreign keys. For
example, if we have 3 ids, and only 2 of them exist in the index
for the nested document type, this will cause nodes to be returned
containing all 3 ids--whereas the unoptimized logic would only return
nodes for the 2 ids that exist in the index. We believe this is an
OK tradeoff to make, but if it becomes a problem, we may want to make
this behavior configurable.