Symlink site artists schema from `elasticgraph new` template.
This reduces duplication, and gives us a strong gurantee that the
example queries on the ElasticGraph website work against the project
template schema. This is important because we plan to include a call
to action on the Query API documentation pages to encourage site
visitors to try out the examples locally by booting the demo app
using our docker image.