Refactor: avoid `index.{md,html}` files. (#333)
`index` is a "special" file name which gets served from `/`. I prefer
not to rely on this; instead it's nice to have files that match the site
structure. For example, `guides/index.md` is not served as a sibling of
the other files under `guides`, and it's nice to avoid giving that
impression.

Note: I've retained the root `index.html` for the home page because I
don't think we can avoid it there!