Respect the configured `index.max_result_window`.
If we submit a query with `size` set to a value larger than `index.max_result_window`,
we get an exception from Elasticsearch/OpenSearch. With the new `size_multiplier`
added in #364, it's quite easy to exceed the `max_result_window` without realizing
it.

To solve this, we can always record the `max_result_window` in our `datastore_config.yaml`.
(Previously, it was only recorded if explicitly configured). Then we can read the value
and respect it when building a datastore query.