Change `HashUtil.fetch*` methods to accept an array instead of a string path.
Profiling has shown that the string splitting/joining in the `HashUtil.fetch*`
methods are a bit of a bottleneck. Using dot-separated string paths results in
excess garbage that needs to be collected, because `string_with_no_dots.split(".")`
produces an extra copy of `string_with_no_dots`.

A direct benchmark also shows that it's faster.

However, in this commit we're not yet benefitting from this optimization:
the `split(".")` has just been moved into the caller. In a later PR, we'll
leverage this by directly working with arrays.