Refactor: introduce `SchemaArtifacts::RuntimeMetadata::GraphQLResolver`.
Previously, `graphql_resolvers_by_name` just contained
`SchemaArtifacts::RuntimeMetadata::Extension` objects.
However, we also want to keep track of `needs_lookahead`
in order to optimize our resolvers (as requestig lookahead
when it is not needed is unnecessarily expensive).

To suppport this, the new `GraphQLResolver` contains an extension
(`resolver_ref`) and `needs_lookahead`.