Improve `script/update_ci_yaml`: tolerate hand-edits to `ci.yaml`.
Previously, all changes to `ci.yaml` had to be made via `script/update_ci_yaml`
(e.g. by editing the template in that file), then running the script. However,
this doesn't playwell with dependabot: it automatically updates the actions
we depend on (which is really useful!) but then that fails the build because
it doesn't also update them in the template we had in `script/update_ci_yaml`.

the new version of the script does not require a template. It just works with
`ci.yaml` as it already exists and does textual updates to the OpenSearch
and Elasticsearch versions based on `tested_datastore_versions.yaml`.