Stop using monkey patching to implement object/field visibility.
As documented in our guiding principles in #470, "We avoid monkey patching."
However, we've been monkey patching `GraphQL::Schema::Object` and
`GraphQL::Schema::Field` because until recently there wasn't an API to
provide custom `visibility?` logic when loading a schema from SDL.
However, an API was recently added to the GraphQL gem[^1] and released
a few weeks ago[^2] which allows us to implement our `visibility?` logic
without any monkey patching. This uses the new API to remove the only
monkey patching in the project, aligning the codebase with one of our
principles.

[^1]: https://github.com/rmosolgo/graphql-ruby/pull/5282
[^2]: https://github.com/rmosolgo/graphql-ruby/blob/master/CHANGELOG.md#2415-19-mar-2025