Use `graphql-c_parser` for faster parsing. (#491)
The graphql-ruby site[^1] mentions that this is a faster parser. Adding
it as a dependency (and loading it when we load the GraphQL gem) ensures
all ElasticGraph users benefit from this optimization. However, if this
causes any problems for any users, we have the option to remove it as a
hard dependency, expecting individual users to add it to their application
bundle (and require it) to benefit from it.

As part of this, I've also removed `spec_support/optimize_graphql`. It
is not compatible with the new spec that verifies that the `CParser` is
assigned as the `::GraphQL.default_parser`. And at this point, I don't
think we need it--between our parallel test runner and the new C parser,
I'm not able to observe a meaningful difference in test suite runtime
with it.

[^1]: https://graphql-ruby.org/language_tools/c_parser.html