Add v0.19.2.2 API docs. (#534)
I released v0.19.2.2 from the `pre-1.0` branch, which means the API docs
tarball was only committed to the `pre-1.0` branch. As a result, the
`v0.19.2.2` docs were not included in the published website.

In addition, use relative links for API doc links. Otherwise, I'm getting a CI failure
due to the new `v0.19.2.2` docs not yet being deployed:

```
For the Links > External check, the following failures were found:

* At /home/runner/work/elasticgraph/elasticgraph/config/site/_site/guides/custom-graphql-resolvers/index.html:578:

  External link https://block.github.io/elasticgraph/api-docs/v0.19.2.2/ElasticGraph/SchemaDefinition/API.html#register_graphql_resolver-instance_method failed (status code 404)

* At /home/runner/work/elasticgraph/elasticgraph/config/site/_site/guides/custom-graphql-resolvers/index.html:617:

  External link https://block.github.io/elasticgraph/api-docs/v0.19.2.2/ElasticGraph/SchemaDefinition/API.html#on_root_query_type-instance_method failed (status code 404)
```

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>