Remove `nested_relationship_resolver_mode`. (#540)
It was added as part of the initial implementation of `NestedRelationshipsSource`
to provide a way to carefully roll it out in production for the first time, but
we've been running in production with it for several weeks and it's been clear
that it's both safe and a vast improvement over the original approach. I don't
think there's any reason to choose the `original` or `comparison` modes and it
simplfies things to drop it.