Remove `TextFilterInput.equal_to_any_of`. (#544)
In my testing, it doesn't actually work. In addition, the Elasticsearch docs[^1] advise against
using a `term` query (which is what `equal_to_any_of` translates into) on a `text` field:

> Avoid using the term query for text fields.
>
> By default, Elasticsearch changes the values of text fields as part of analysis. This can make finding exact matches for text field values difficult.
>
> To search text field values, use the match query instead.

Notably, `textField: {not: {equalToAnyOf: [null]}}` does work (because it translates into
an `exists` query, not a `terms` query). However, we don't want to leave a filtering predicate
in place that works for only one argument value and not for all others.

[^1]: https://www.elastic.co/docs/reference/query-languages/query-dsl/query-dsl-term-query