Prefer using `filter` over `must` in boolean queries.
As explained by the Elasticsearch docs[^1], `filter` and `must` are nearly identical. The
difference has to do with scoring and caching:

* Clauses under `must` contribute to the relevance score, and are not considered for caching.
* Clauses under `filter` ignore relevance scoring and are considered for caching.

We don't do anything with the relevance score (for example, we _always_ sort by a concrete field,
and don't expose an API to interact with the score), so our current use of `must` isn't doing
anything for us, other than preventing Elasticsearch/OpenSearch from applying caching.

Switching to `filter` should be a bit more performant, with no difference in observable query behavior.

[^1]: https://www.elastic.co/docs/reference/query-languages/query-dsl/query-dsl-bool-query