Introduce `internal_filters` vs `client_filters` on `DatastoreQuery`.
Previously, we just had `filters`. That's been replaced by three attributes:

* `client_filters` for filters that were directly provided in the GraphQL query by the client.
* `internal_filters` for filters added internally by ElasticGraph or a query interceptor.
* `all_filters` for the combined set of `client_filters` and `internal_filters`.

This split is a necessary change to support search highlighting as planned in #561. For the
`all_highlights` field we'll be adding to edge types, we only want to include highlights that
actually came from the client-specified filters. Filters added internally by ElasticGraph or
a query interceptor should not produce highlights we return to the client as those filters
are hidden/private/internal.

To faciliate that, we'll be using `highlight_query`, which allows us to specify a query for
the purposes of highlighting. We'll use `all_filters` for the main search query but only
`client_filters` for the `highlight_query`.

Note: this is a bit of a breaking change for ElasticGraph projects that use
`elasticgraph-query_interceptor` to add internal filters (as the query interceptors
will have to migrate from `filters` to updating `internal_filters`) but this is the
ideal time to make that change, as our next release will be 1.0.0.