Bugfix: correctly handle graphql-only fields defined with `name_in_index`.
This can be used, for example, when migrating a field in an index to a
new GraphQL field name. Previously, we were not able to correctly
determine the mapping of the field in this case (since it's graphql-only).
With this fix, we now use the mapping of the indexing field with the same
name. This ensures we treat `text` and `nested` fields correctly, for
example.