Breaking change: improve schema definition resolver API.
Instead of an assignment API that looks like this:

```ruby
schema.object_type "MyType" do |t|
  t.default_graphql_resolver = :my_resolver

  t.field "field1", "String"
  t.field "field2", "String" do |f|
    f.resolver = :my_other_resolver
  end
end
```

It now looks like this:

```ruby
schema.object_type "MyType" do |t|
  t.resolve_fields_with :my_resolver

  t.field "field1", "String"
  t.field "field2", "String" do |f|
    f.resolve_with :my_other_resolver
  end
end
```

This is more in keeping with our other schema definition APIs (e.g.
`f.documentation "docstring" instead of `f.documentation = "docstring"`)
and paves the way for these APIs to accept additional arguments.

While this is a breaking change, now is an ideal time to break things,
in the runup to the 1.0.0 release.