Provide `apollo_entity_ref_field` to expose an id field as an entity ref.
When plugging into an Apollo supergraph, it's quite useful to be able to
expoes "entity reference" types (e.g. a type with just an id field, that
refers to an entity owned by another subgraph) rather than just exposing
an id field. This allows Apollo to federate the type.

This adds two new APIs to support this:

* `apollo_entity_ref_field`
* `apollo_entity_ref_paginated_collection_field`

The latter exposes a list of ids as a paginated relay connection.