fix: make `count` filters on a sub-aggregation work correctly. (#660)
A `count` filter translates into an OpenSearch/Elasticsearch query
filter against a special `__counts` field which is a flat map containing
the counts of each list under the current document context. Our
special handling for it worked everywhere except on a sub-aggregation.
This fixes it by correctly converting the field path to a nested
field path.

Fixes #635.