fix: fail with a clear error when `nested` mapping type is improperly used. (#701)
* fix: fail with a clear error when `nested` mapping type is improperly used.

The `nested` mapping type is only valid on a list-of-objects field
(or a `paginated_collection_field` with an element type that is an object).

If `nested` is used on a list-of-scalars field, confusing GraphQL schema
errors could occur, such as:

```
GraphQL::Schema::InvalidDocumentError: Type "TeamStringSubAggregationConnection" not found in document.
```

This avoids those confusing errors.

* Update elasticgraph-schema_definition/lib/elastic_graph/schema_definition/schema_elements/field.rb

Co-authored-by: Fasih Awan <73565509+fasihawan@users.noreply.github.com>

---------

Co-authored-by: Fasih Awan <73565509+fasihawan@users.noreply.github.com>