[chore] Fix CI build: deal with new `America/Coyhaique` timezone.
Our CI builds have started failing on this test. For now we'll ignore
this new timezone (as we don't expect it to be widely supported on all
Elasticsearch/OpenSearch versions we support). In the future, we can
add it to `valid_time_zones.rb` (via `script/dump_time_zones`) and
remove its exclusion here.