feat: migrate `Support::Logger::Config` to use `ElasticGraph::Config`.
This improves the validation of the logger config and participates in
our new config system which will be used to drive our documentation.

In addition, this provides defaults for all `logger` config values so
that ElasticGraph can be booted without any `logger` config. Previously,
`device` had no default and was required. Now it defaults to `stdout`.