Introduce `ElasticGraph::Config`.
This defines a configuration class which uses JSON schema to validate
the provided configuration, removing the need to perform ad-hoc config
validation like we were before. In addition, the JSON schema will allow
us to improve our documentation as we can merge the JSON schemas from
all configuration classes to provide the overall config schema of
ElasticGraph in our docs. (That will come in a later PR).