Provide a configuration settings reference.
This adds a new doc to the versioned YARD docs. This new file contains:

- An extensive config settings example.
- The JSON schema for config settings.

In the future, I want to include similar content as a guide on the main
part of the website, but config settings may change between ElatsicGraph
versions and its nice to provide access to past versions in our YARD
docs.