Generate an overall JSON schema for all config.
This schema can be used to validate an ElasticGraph config file as well
as in local tooling (such as IDEs). For now, we use it in the common
project specs to validate all settings files. We'll do more with it
in later PRs.

In addition, this hooks into the CI build to validate that the JSON
schema is kept up-to-date.