Move `ElasticGraph::Config` to `ElasticGraph::Support::Config`.
Generally speaking, it's a good practice to nest all of a gem's classes
under the gem's namespace. This helps avoid naming collisions and conflicts.
While we have made a couple of exceptions in `elasticgraph-support`
(e.g. `elasticgraph-support/lib/elastic_graph/version.rb` defines `ElasticGraph::VERSION`
since it defines a version for ALL of ElasticGraph), this config class doesn't
warrant that treatment. It's a supporting class, not a definition of all of the
config for ElasticGraph.