chore: Improve CI build by validating the release workflow.
The release workflow has broken a couple times recently, and had to be
fixed up via #759 and #797. The fact that the workflow is only used when
we cut a release means that it does not run super often (and certainly
not part of the CI build for each PR!).

To help avoid breaking it in the future, I've added a new script,
`script/validate_release_workflow`, which validates the brittle part
of the release workflow, and I've integrated it into the CI build
so that a build will fail if a change would break the release workflow.