chore: Fix YAML syntax highlighting at the website.
The syntax highlighter we use, `rouge`, does not handle quoted keys
(such as `"$schema"`) correctly--it's been highlighting them as having
a syntax error. I dug into the issue in rouge, created a fix, and have
a PR against rouge opened with the fix. This pulls in the fix from my
fork so we can benefit from it now. Once my fix is merged and released,
we can switch to the released version.