Refactor: stop exposing `indices` when we only support one. (#819)
Originally, we thought we might support multiple but there's never been
a need and it's annoying to always have to deal with `indices` as a list
when we only have one (or none).

As part of this, we've changed the behavior slightly for when `index`
is called more than once: previously, we replaced the existing definition.
Now we raise an error as we should.